# Conflicts:
#	app-service/src/main/java/in/gov/abdm/abha/enrollment/controller/AuthController.java

+++ db-service/src/main/java/in/gov/abdm/abha/enrollmentdb/model/dependentaccountrelationship/DependentAccountRelationship.java
@@ -20,13 +20,13 @@ import java.time.LocalDateTime;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class DependentAccountRelationship implements Persistable<BigInteger> {
+public class DependentAccountRelationship implements Persistable<Long> {
 
     /**
      *  it is Id and Primary key
      */
     @Id
-    private BigInteger id;
+    private Long id;
 
     /**
      * it is a parent_health_id_number it is a Foreign key
